src/requirements.txt
python-etcd
docker-py
